+++ src/main/java/dev/virtuoso/ims/repository/interest/LoanInterestParamTypeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LoanInterestParamTypeRepository extends JpaRepository<LoanInterestParamType, Long> {
+public interface LoanInterestParamTypeRepository extends JpaRepository<LoanInterestParamType, Integer> {
 }
